chore: move getTemplatesById to /template
