+++ src/dataset/domain/hooks/useGetTemplatesByCollectionId.ts
@@ -2,8 +2,8 @@ import { useCallback, useEffect, useState } from 'react'
 import { ReadError } from '@iqss/dataverse-client-javascript'
 import { JSDataverseReadErrorHandler } from '@/shared/helpers/JSDataverseReadErrorHandler'
 import { TemplateRepository } from '@/templates/domain/repositories/TemplateRepository'
-import { Template } from '../models/DatasetTemplate'
-import { getTemplatesByCollectionId } from '../useCases/getTemplatesByCollectionId'
+import { Template } from '@/templates/domain/models/Template'
+import { getTemplatesByCollectionId } from '@/templates/domain/useCases/getTemplatesByCollectionId'
 
 interface useGetTemplatesByCollectionIdProps {
   templateRepository: TemplateRepository
