+++ tests/component/dataset/domain/models/DatasetTemplateMother.ts
@@ -1,5 +1,5 @@
 import { faker } from '@faker-js/faker'
-import { Template } from '@/dataset/domain/models/DatasetTemplate'
+import { Template } from '@/templates/domain/models/Template'
 
 export class DatasetTemplateMother {
   static createMany(count: number, props?: Partial<Template>): Template[] {
